create component GridItem and fixed style UI in popularBlock and Prooduct select color
